Locking Down Mutations (Posts and Comments)

+++ graphql-prisma/src/resolvers/Mutation.js
@@ -85,17 +85,17 @@ const Mutation = {
       info
     );
   },
-  async deletePost(parent, args, { prisma,request }, info) {
-     const userId = getUserId(request);
-     const postExists=await prisma.exists.Post({
-       id:args.id,
-       author:{
-         id:userId
-       }
-     })
-     if (!postExists) {
-       throw new Error("Unable to delete post")
-     }
+  async deletePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+    if (!postExists) {
+      throw new Error("Unable to delete post");
+    }
     return prisma.mutation.deletePost(
       {
         where: {
@@ -105,7 +105,17 @@ const Mutation = {
       info
     );
   },
-  updatePost(parent, args, { prisma }, info) {
+  async updatePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+    if (!postExists) {
+      throw new Error("Unable to update post");
+    }
     return prisma.mutation.updatePost(
       {
         where: {
@@ -116,14 +126,15 @@ const Mutation = {
       info
     );
   },
-  createComment(parent, args, { prisma }, info) {
+  createComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
     return prisma.mutation.createComment(
       {
         data: {
           text: args.data.text,
           author: {
             connect: {
-              id: args.data.author
+              id: userId
             }
           },
           post: {
@@ -136,7 +147,17 @@ const Mutation = {
       info
     );
   },
-  deleteComment(parent, args, { prisma }, info) {
+  async deleteComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+    if (!commentExists) {
+      throw new Error("Unable to delete the comment");
+    }
     return prisma.mutation.deleteComment(
       {
         where: {
@@ -146,7 +167,17 @@ const Mutation = {
       info
     );
   },
-  updateComment(parent, args, { prisma }, info) {
+  async updateComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+    if (!commentExists) {
+      throw new Error("Unable to update the comment");
+    }
     return prisma.mutation.updateComment(
       {
         where: {
